Fix b2b example to match skeleton
