+++ examples/b2b/app/components/B2BLocationProvider.tsx
@@ -1,6 +1,6 @@
 import {createContext, useContext, useEffect, useState, useMemo} from 'react';
 import {useFetcher} from '@remix-run/react';
-import {CustomerCompany} from '../root';
+import {type CustomerCompany} from '~/root';
 
 export type B2BLocationContextValue = {
   company?: CustomerCompany;
