+++ examples/b2b/app/routes/products.$handle.tsx
@@ -1,35 +1,17 @@
 import {Suspense} from 'react';
 import {defer, redirect, type LoaderFunctionArgs} from '@shopify/remix-oxygen';
+import {Await, useLoaderData, type MetaFunction} from '@remix-run/react';
+import type {ProductFragment} from 'storefrontapi.generated';
 import {
-  Await,
-  Link,
-  useLoaderData,
-  type MetaFunction,
-  type FetcherWithComponents,
-} from '@remix-run/react';
-import type {
-  ProductFragment,
-  ProductVariantsQuery,
-  ProductVariantFragment,
-} from 'storefrontapi.generated';
-
-import {
-  Image,
-  Money,
-  VariantSelector,
-  type VariantOption,
   getSelectedProductOptions,
-  CartForm,
   Analytics,
-  type CartViewPayload,
-  useAnalytics,
+  useOptimisticVariant,
 } from '@shopify/hydrogen';
-import type {
-  CartLineInput,
-  SelectedOption,
-} from '@shopify/hydrogen/storefront-api-types';
+import type {SelectedOption} from '@shopify/hydrogen/storefront-api-types';
 import {getVariantUrl} from '~/lib/variants';
-import {useAside} from '~/components/Aside';
+import {ProductPrice} from '~/components/ProductPrice';
+import {ProductImage} from '~/components/ProductImage';
+import {ProductForm} from '~/components/ProductForm';
 import {QuantityRules, hasQuantityRules} from '~/components/QuantityRules';
 import {PriceBreaks} from '~/components/PriceBreaks';
 
@@ -37,49 +19,24 @@ export const meta: MetaFunction<typeof loader> = ({data}) => {
   return [{title: `Hydrogen | ${data?.product.title ?? ''}`}];
 };
 
+/***********************************************/
+/**********  EXAMPLE UPDATE STARTS  ************/
+type BuyerVariables = {
+  buyer: {
+    companyLocationId: string;
+    customerAccessToken: string;
+  };
+} | {};
+/**********   EXAMPLE UPDATE END   *************/
+/***********************************************/
+
 export async function loader(args: LoaderFunctionArgs) {
-  // Start fetching non-critical data without blocking time to first byte
-  const deferredData = loadDeferredData(args);
-
-  // Await the critical data required to render initial state of the page
-  const criticalData = await loadCriticalData(args);
-
-  return defer({...deferredData, ...criticalData});
-}
-
-/**
- * Load data necessary for rendering content above the fold. This is the critical data
- * needed to render the page. If it's unavailable, the whole page should 400 or 500 error.
- */
-async function loadCriticalData({
-  context,
-  params,
-  request,
-}: LoaderFunctionArgs) {
-  const {handle} = params;
-  const {storefront, customerAccount} = context;
-
-  const selectedOptions = getSelectedProductOptions(request).filter(
-    (option) =>
-      // Filter out Shopify predictive search query params
-      !option.name.startsWith('_sid') &&
-      !option.name.startsWith('_pos') &&
-      !option.name.startsWith('_psq') &&
-      !option.name.startsWith('_ss') &&
-      !option.name.startsWith('_v') &&
-      // Filter out third party tracking params
-      !option.name.startsWith('fbclid'),
-  );
-
-  if (!handle) {
-    throw new Error('Expected product handle to be defined');
-  }
 
   /***********************************************/
   /**********  EXAMPLE UPDATE STARTS  ************/
-  const buyer = await customerAccount.UNSTABLE_getBuyer();
+  const buyer = await args.context.customerAccount.UNSTABLE_getBuyer();
 
-  const buyerVariables =
+  const buyerVariables: BuyerVariables =
     buyer?.companyLocationId && buyer?.customerAccessToken
       ? {
           buyer: {
@@ -89,13 +46,48 @@ async function loadCriticalData({
         }
       : {};
 
-  // await the query for the critical product data
-  const {product} = await storefront.query(PRODUCT_QUERY, {
-    variables: {handle, selectedOptions, ...buyerVariables},
-    cache: storefront.CacheNone(),
-  });
+  // Start fetching non-critical data without blocking time to first byte
+  const deferredData = loadDeferredData(args, buyerVariables);
+
+  // Await the critical data required to render initial state of the page
+  const criticalData = await loadCriticalData(args, buyerVariables);
   /**********   EXAMPLE UPDATE END   *************/
   /***********************************************/
+
+  return defer({...deferredData, ...criticalData});
+}
+
+/**
+ * Load data necessary for rendering content above the fold. This is the critical data
+ * needed to render the page. If it's unavailable, the whole page should 400 or 500 error.
+ */
+/***********************************************/
+/**********  EXAMPLE UPDATE STARTS  ************/
+async function loadCriticalData({
+  context,
+  params,
+  request,
+}: LoaderFunctionArgs, buyerVariables: BuyerVariables) {
+  const {handle} = params;
+  const {storefront} = context;
+
+  if (!handle) {
+    throw new Error('Expected product handle to be defined');
+  }
+
+  const [{product}] = await Promise.all([
+    storefront.query(PRODUCT_QUERY, {
+      variables: {
+        handle,
+        selectedOptions: getSelectedProductOptions(request),
+        ...buyerVariables,
+      },
+      cache: storefront.CacheNone(),
+    }),
+    // Add other queries here, so that they are loaded in parallel
+  ]);
+/**********   EXAMPLE UPDATE END   *************/
+/***********************************************/
 
   if (!product?.id) {
     throw new Response(null, {status: 404});
@@ -119,21 +111,9 @@ async function loadCriticalData({
     }
   }
 
-  // In order to show which variants are available in the UI, we need to query
-  // all of them. But there might be a *lot*, so instead separate the variants
-  // into it's own separate query that is deferred. So there's a brief moment
-  // where variant options might show as available when they're not, but after
-  // this deffered query resolves, the UI will update.
-  /***********************************************/
-  /**********  EXAMPLE UPDATE STARTS  ************/
-  const variants = storefront.query(VARIANTS_QUERY, {
-    variables: {handle, ...buyerVariables},
-    cache: storefront.CacheNone(),
-  });
-  /**********   EXAMPLE UPDATE END   *************/
-  /***********************************************/
-
-  return {product, variants};
+  return {
+    product,
+  };
 }
 
 /**
@@ -141,9 +121,33 @@ async function loadCriticalData({
  * fetched after the initial page load. If it's unavailable, the page should still 200.
  * Make sure to not throw any errors here, as it will cause the page to 500.
  */
-function loadDeferredData({context}: LoaderFunctionArgs) {
-  return {};
+  /***********************************************/
+  /**********  EXAMPLE UPDATE STARTS  ************/
+function loadDeferredData({context, params}: LoaderFunctionArgs, buyerVariables: BuyerVariables) {
+  const {storefront} = context;
+
+  // In order to show which variants are available in the UI, we need to query
+  // all of them. But there might be a *lot*, so instead separate the variants
+  // into it's own separate query that is deferred. So there's a brief moment
+  // where variant options might show as available when they're not, but after
+  // this deferred query resolves, the UI will update.
+  const variants = context.storefront
+    .query(VARIANTS_QUERY, {
+      variables: {handle: params.handle!, ...buyerVariables},
+      cache: storefront.CacheNone(),
+    })
+    .catch((error) => {
+      // Log query errors, but don't throw them so the page can still render
+      console.error(error);
+      return null;
+    });
+
+  return {
+    variants,
+  };
 }
+/**********   EXAMPLE UPDATE END   *************/
+/***********************************************/
 
 function redirectToFirstVariant({
   product,
@@ -170,15 +174,81 @@ function redirectToFirstVariant({
 
 export default function Product() {
   const {product, variants} = useLoaderData<typeof loader>();
-  const {selectedVariant} = product;
+  const selectedVariant = useOptimisticVariant(
+    product.selectedVariant,
+    variants,
+  );
+
+  const {title, descriptionHtml} = product;
+
   return (
     <div className="product">
       <ProductImage image={selectedVariant?.image} />
-      <ProductMain
-        selectedVariant={selectedVariant}
-        product={product}
-        variants={variants}
-      />
+      <div className="product-main">
+        <h1>{title}</h1>
+        <ProductPrice
+          price={selectedVariant?.price}
+          compareAtPrice={selectedVariant?.compareAtPrice}
+        />
+        <br />
+        <Suspense
+          fallback={
+            <ProductForm
+              product={product}
+              selectedVariant={selectedVariant}
+              variants={[]}
+            />
+          }
+        >
+          <Await
+            errorElement="There was a problem loading product variants"
+            resolve={variants}
+          >
+            {(data) => (
+              <ProductForm
+                product={product}
+                selectedVariant={selectedVariant}
+                variants={data?.product?.variants.nodes || []}
+                /***********************************************/
+                /**********  EXAMPLE UPDATE STARTS  ************/
+                quantity={selectedVariant?.quantityRule?.increment || 1}
+                /**********   EXAMPLE UPDATE END   ************/
+                /***********************************************/
+              />
+            )}
+          </Await>
+        </Suspense>
+        <br />
+        {
+          /***********************************************/
+          /**********  EXAMPLE UPDATE STARTS  ************/
+          hasQuantityRules(selectedVariant?.quantityRule) ? (
+            <QuantityRules
+              maximum={selectedVariant?.quantityRule.maximum}
+              minimum={selectedVariant?.quantityRule.minimum}
+              increment={selectedVariant?.quantityRule.increment}
+            />
+          ) : null
+        }
+        <br />
+        {
+          selectedVariant?.quantityPriceBreaks?.nodes &&
+          selectedVariant?.quantityPriceBreaks?.nodes?.length > 0 ? (
+            <PriceBreaks
+              priceBreaks={selectedVariant?.quantityPriceBreaks?.nodes}
+            />
+          ) : null
+          /**********   EXAMPLE UPDATE END   ************/
+          /***********************************************/
+        }
+        <br />
+        <p>
+          <strong>Description</strong>
+        </p>
+        <br />
+        <div dangerouslySetInnerHTML={{__html: descriptionHtml}} />
+        <br />
+      </div>
       <Analytics.ProductView
         data={{
           products: [
@@ -195,250 +265,6 @@ export default function Product() {
         }}
       />
     </div>
-  );
-}
-
-function ProductImage({image}: {image: ProductVariantFragment['image']}) {
-  if (!image) {
-    return <div className="product-image" />;
-  }
-  return (
-    <div className="product-image">
-      <Image
-        alt={image.altText || 'Product Image'}
-        aspectRatio="1/1"
-        data={image}
-        key={image.id}
-        sizes="(min-width: 45em) 50vw, 100vw"
-      />
-    </div>
-  );
-}
-
-function ProductMain({
-  selectedVariant,
-  product,
-  variants,
-}: {
-  product: ProductFragment;
-  selectedVariant: ProductFragment['selectedVariant'];
-  variants: Promise<ProductVariantsQuery>;
-}) {
-  const {title, descriptionHtml} = product;
-
-  return (
-    <div className="product-main">
-      <h1>{title}</h1>
-      <ProductPrice selectedVariant={selectedVariant} />
-      <br />
-      <Suspense
-        fallback={
-          <ProductForm
-            product={product}
-            selectedVariant={selectedVariant}
-            variants={[]}
-            quantity={1}
-          />
-        }
-      >
-        <Await
-          errorElement="There was a problem loading product variants"
-          resolve={variants}
-        >
-          {(data) => (
-            <ProductForm
-              product={product}
-              selectedVariant={selectedVariant}
-              variants={data.product?.variants.nodes || []}
-              /***********************************************/
-              /**********  EXAMPLE UPDATE STARTS  ************/
-              quantity={selectedVariant?.quantityRule?.increment || 1}
-              /**********   EXAMPLE UPDATE END   ************/
-              /***********************************************/
-            />
-          )}
-        </Await>
-      </Suspense>
-      <br />
-      {
-        /***********************************************/
-        /**********  EXAMPLE UPDATE STARTS  ************/
-        hasQuantityRules(selectedVariant?.quantityRule) ? (
-          <QuantityRules
-            maximum={selectedVariant?.quantityRule.maximum}
-            minimum={selectedVariant?.quantityRule.minimum}
-            increment={selectedVariant?.quantityRule.increment}
-          />
-        ) : null
-      }
-      <br />
-      {
-        selectedVariant?.quantityPriceBreaks?.nodes &&
-        selectedVariant?.quantityPriceBreaks?.nodes?.length > 0 ? (
-          <PriceBreaks
-            priceBreaks={selectedVariant?.quantityPriceBreaks?.nodes}
-          />
-        ) : null
-        /**********   EXAMPLE UPDATE END   ************/
-        /***********************************************/
-      }
-      <br />
-      <p>
-        <strong>Description</strong>
-      </p>
-      <br />
-      <div dangerouslySetInnerHTML={{__html: descriptionHtml}} />
-      <br />
-    </div>
-  );
-}
-
-function ProductPrice({
-  selectedVariant,
-}: {
-  selectedVariant: ProductFragment['selectedVariant'];
-}) {
-  return (
-    <div className="product-price">
-      {selectedVariant?.compareAtPrice ? (
-        <>
-          <p>Sale</p>
-          <br />
-          <div className="product-price-on-sale">
-            {selectedVariant ? <Money data={selectedVariant.price} /> : null}
-            <s>
-              <Money data={selectedVariant.compareAtPrice} />
-            </s>
-          </div>
-        </>
-      ) : (
-        selectedVariant?.price && <Money data={selectedVariant?.price} />
-      )}
-    </div>
-  );
-}
-
-/***********************************************/
-/**********  EXAMPLE UPDATE STARTS  ************/
-function ProductForm({
-  product,
-  selectedVariant,
-  variants,
-  quantity,
-}: {
-  product: ProductFragment;
-  selectedVariant: ProductFragment['selectedVariant'];
-  variants: Array<ProductVariantFragment>;
-  quantity: number;
-}) {
-  /**********   EXAMPLE UPDATE END   ************/
-  /***********************************************/
-  const {open} = useAside();
-  const {publish, shop, cart, prevCart} = useAnalytics();
-  return (
-    <div className="product-form">
-      <VariantSelector
-        handle={product.handle}
-        // Only show options with more than one value
-        options={product.options.filter((option) => option.values.length > 1)}
-        variants={variants}
-      >
-        {({option}) => <ProductOptions key={option.name} option={option} />}
-      </VariantSelector>
-      <br />
-      <AddToCartButton
-        disabled={!selectedVariant || !selectedVariant.availableForSale}
-        onClick={() => {
-          open('cart');
-          publish('cart_viewed', {
-            cart,
-            prevCart,
-            shop,
-            url: window.location.href || '',
-          } as CartViewPayload);
-        }}
-        lines={
-          selectedVariant
-            ? [
-                {
-                  merchandiseId: selectedVariant.id,
-                  /***********************************************/
-                  /**********  EXAMPLE UPDATE STARTS  ************/
-                  quantity,
-                  /**********   EXAMPLE UPDATE END   ************/
-                  /***********************************************/
-                },
-              ]
-            : []
-        }
-      >
-        {selectedVariant?.availableForSale ? 'Add to cart' : 'Sold out'}
-      </AddToCartButton>
-    </div>
-  );
-}
-
-function ProductOptions({option}: {option: VariantOption}) {
-  return (
-    <div className="product-options" key={option.name}>
-      <h5>{option.name}</h5>
-      <div className="product-options-grid">
-        {option.values.map(({value, isAvailable, isActive, to}) => {
-          return (
-            <Link
-              className="product-options-item"
-              key={option.name + value}
-              prefetch="intent"
-              preventScrollReset
-              replace
-              to={to}
-              style={{
-                border: isActive ? '1px solid black' : '1px solid transparent',
-                opacity: isAvailable ? 1 : 0.3,
-              }}
-            >
-              {value}
-            </Link>
-          );
-        })}
-      </div>
-      <br />
-    </div>
-  );
-}
-
-function AddToCartButton({
-  analytics,
-  children,
-  disabled,
-  lines,
-  onClick,
-}: {
-  analytics?: unknown;
-  children: React.ReactNode;
-  disabled?: boolean;
-  lines: CartLineInput[];
-  onClick?: () => void;
-}) {
-  return (
-    <CartForm route="/cart" inputs={{lines}} action={CartForm.ACTIONS.LinesAdd}>
-      {(fetcher: FetcherWithComponents<any>) => (
-        <>
-          <input
-            name="analytics"
-            type="hidden"
-            value={JSON.stringify(analytics)}
-          />
-          <button
-            type="submit"
-            onClick={onClick}
-            disabled={disabled ?? fetcher.state !== 'idle'}
-          >
-            {children}
-          </button>
-        </>
-      )}
-    </CartForm>
   );
 }
