+++ examples/b2b/app/styles/app.css
@@ -112,14 +112,34 @@ aside li {
   width: 100%;
 }
 
-.overlay:target {
+.overlay.expanded {
   opacity: 1;
   pointer-events: auto;
   visibility: visible;
 }
 /* reveal aside */
-.overlay:target aside {
+.overlay.expanded aside {
   transform: translateX(calc(var(--aside-width) * -1));
+}
+
+button.reset {
+  border: 0;
+  background: inherit;
+  font-size: inherit;
+}
+
+button.reset > * {
+  margin: 0;
+}
+
+button.reset:not(:has(> *)) {
+  height: 1.5rem;
+  line-height: 1.5rem;
+}
+
+button.reset:hover:not(:has(> *)) {
+  text-decoration: underline;
+  cursor: pointer;
 }
 
 /*
@@ -298,8 +318,6 @@ aside li {
   align-items: center;
   gap: 1rem;
 }
-
-
 
 /*
 * --------------------------------------------------
@@ -498,23 +516,23 @@ aside li {
 * --------------------------------------------------
 */
 .modal {
-  position: fixed; 
+  position: fixed;
   display: flex;
-  z-index: 1; 
+  z-index: 1;
   left: 0;
   top: 0;
-  width: 100%; 
-  height: 100%; 
-  background-color: rgba(0,0,0,0.4); 
+  width: 100%;
+  height: 100%;
+  background-color: rgba(0,0,0,0.4);
 }
 
 .modal-content {
   background-color: #fefefe;
   overflow: auto;
-  margin: auto; 
+  margin: auto;
   border: 1px solid #888;
   border-radius: 4px;
-  width: fit-content; 
+  width: fit-content;
   max-height: 80%;
   padding: 20px
 }
